+++ src/onegov.org/onegov/org/models/__init__.py
@@ -5,21 +5,23 @@ from onegov.org.models.directory import ExtendedDirectory
 from onegov.org.models.directory import ExtendedDirectoryEntry
 from onegov.org.models.editor import Editor
 from onegov.org.models.export import Export, ExportCollection
-from onegov.org.models.file import (
-    GeneralFile,
-    GeneralFileCollection,
-    ImageFile,
-    ImageFileCollection,
-    ImageSet,
-    ImageSetCollection,
-)
+from onegov.org.models.extensions import ContactExtension
+from onegov.org.models.extensions import ContentExtension
+from onegov.org.models.extensions import CoordinatesExtension
+from onegov.org.models.extensions import HiddenFromPublicExtension
+from onegov.org.models.extensions import PersonLinkExtension
+from onegov.org.models.extensions import VisibleOnHomepageExtension
+from onegov.org.models.file import GeneralFile
+from onegov.org.models.file import GeneralFileCollection
+from onegov.org.models.file import ImageFile
+from onegov.org.models.file import ImageFileCollection
+from onegov.org.models.file import ImageSet
+from onegov.org.models.file import ImageSetCollection
 from onegov.org.models.form import BuiltinFormDefinition, CustomFormDefinition
-from onegov.org.models.legacy_file import (
-    LegacyFile,
-    LegacyImage,
-    LegacyFileCollection,
-    LegacyImageCollection,
-)
+from onegov.org.models.legacy_file import LegacyFile
+from onegov.org.models.legacy_file import LegacyFileCollection
+from onegov.org.models.legacy_file import LegacyImage
+from onegov.org.models.legacy_file import LegacyImageCollection
 from onegov.org.models.message import DirectoryMessage
 from onegov.org.models.message import EventMessage
 from onegov.org.models.message import PaymentMessage
@@ -30,28 +32,17 @@ from onegov.org.models.message import TicketNote
 from onegov.org.models.organisation import Organisation
 from onegov.org.models.page import AtoZPages, News, Topic
 from onegov.org.models.page_move import PageMove
-from onegov.org.models.person_move import (
-    FormPersonMove,
-    PagePersonMove,
-    PersonMove,
-    ResourcePersonMove,
-)
-from onegov.org.models.recipient import (
-    ResourceRecipient,
-    ResourceRecipientCollection
-)
+from onegov.org.models.person_move import FormPersonMove
+from onegov.org.models.person_move import PagePersonMove
+from onegov.org.models.person_move import PersonMove
+from onegov.org.models.person_move import ResourcePersonMove
+from onegov.org.models.publication import PublicationCollection
+from onegov.org.models.recipient import ResourceRecipient
+from onegov.org.models.recipient import ResourceRecipientCollection
 from onegov.org.models.resource import DaypassResource
 from onegov.org.models.search import Search
 from onegov.org.models.sitecollection import SiteCollection
 from onegov.org.models.traitinfo import TraitInfo
-from onegov.org.models.extensions import (
-    ContactExtension,
-    ContentExtension,
-    CoordinatesExtension,
-    HiddenFromPublicExtension,
-    PersonLinkExtension,
-    VisibleOnHomepageExtension,
-)
 
 __all__ = [
     'AtoZ',
@@ -90,6 +81,7 @@ __all__ = [
     'PaymentMessage',
     'PersonLinkExtension',
     'PersonMove',
+    'PublicationCollection',
     'ReservationMessage',
     'ResourcePersonMove',
     'ResourceRecipient',